requirements.txt
Django>=5.2
Pillow>=10.0.0
python-dateutil>=2.8.2
whitenoise>=6.5.0
psycopg2-binary>=2.9.0
redis>=4.5.0
gunicorn>=21.0.0
python-dotenv>=1.0.0